pathvirt.cpp: Optimize clfind() to return path ptr and index
This patch adds a minor optimization to the clfind() function --
it now returns the pointer to the location of a path prefix along
with the index in a colon separated list of path prefixes.

With the earlier implementation, we were traversing the old list
twice: first time to find the index of a given path; and the second
time to find the pointer to the path in order to figure out its size.